requirements.txt
tensorflow==1.15.4
azure-storage-blob==2.1.0
azure-iot-hub==2.4.0
